Move do_* methods into the related extrinsic module (bittensor/core/extrinsics/serving.py)
